Example full_app and full_app tests refactor
